Python Templates/requirements.txt
nextcord
load_dotenv
aiosqlite
datetime
asyncio